requirements.txt
Keras==2.1.4
numpy==1.16.5
tensorflow_gpu==1.15.0
pandas==0.25.3
scikit-learn==0.22
coverage
pipreqs
psutil
